Moved loader into its own file
