+++ tgui/packages/tgui/interfaces/Input/Loader.tsx
@@ -1,0 +1,15 @@
+import { Box } from '../../components';
+import { clamp01 } from 'common/math';
+
+/** Progress bar for timed input prompts */
+export const Loader = props => {
+  const { value } = props;
+
+  return (
+    <div className="AlertModal__Loader">
+      <Box
+        className="AlertModal__LoaderProgress"
+        style={{ width: clamp01(value) * 100 + '%' }} />
+    </div>
+  );
+};
